backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
librosa
soundfile
numpy
scipy
matplotlib
umap-learn
scikit-learn
pillow
python-dotenv
